Check for file/fiolder existence

+++ test.py
@@ -152,17 +152,21 @@ def get_path_to_folder_for_images():
 
 
 def remove_file(t_path):
-    try:
-        os.remove(t_path)
-    except Exception as err:
-        print("Error during file remove: {0}".format(err))
+    if os.path.exist(t_path):
+        try:
+            os.remove(t_path)
+        except Exception as err:
+            print("Error during file remove: {0}".format(err))
+            return
 
 
 def remove_folder(t_path):
-    try:
-        shutil.rmtree(t_path, True)
-    except Exception as err:
-        print("Error during folder remove: {0}".format(err))
+    if os.path.exist(t_path):
+        try:
+            shutil.rmtree(t_path, True)
+        except Exception as err:
+            print("Error during folder remove: {0}".format(err))
+            return
 
 
 if __name__ == '__main__':
